refactor: remove dead v12 internals
