+++ packages/format-explorer/src/utils.ts
@@ -1,6 +1,3 @@
-import { customRef } from 'vue'
-import type { Ref } from 'vue'
-
 export function debounce<T extends (...args: any[]) => any>(fn: T, delay: number = 300): T {
   let prevTimer: number | null = null
   return ((...args: any[]) => {
@@ -12,23 +9,4 @@ export function debounce<T extends (...args: any[]) => any>(fn: T, delay: number
       prevTimer = null
     }, delay)
   }) as any
-}
-
-export function useDebouncedRef<T>(value: T, delay = 200): Ref<T> {
-  let timeout: ReturnType<typeof setTimeout>
-  return customRef<T>((track, trigger) => {
-    return {
-      get() {
-        track()
-        return value
-      },
-      set(newValue: T) {
-        clearTimeout(timeout)
-        timeout = setTimeout(() => {
-          value = newValue
-          trigger()
-        }, delay)
-      }
-    }
-  })
 }
