+++ packages/vue-i18n-core/src/components/DatetimeFormat.ts
@@ -41,8 +41,7 @@ export const DatetimeFormatImpl: ComponentOptions<DatetimeFormatProps> =
       const i18n =
         props.i18n ||
         (useI18n({
-          useScope: props.scope as 'global' | 'parent',
-          __useComponent: true
+          useScope: props.scope as 'global' | 'parent'
         }) as unknown as Composer & ComposerInternal)
 
       return renderFormatter<
