+++ packages/vue-i18n-core/src/components/NumberFormat.ts
@@ -40,8 +40,7 @@ export const NumberFormatImpl: ComponentOptions<NumberFormatProps> = /* #__PURE_
       const i18n =
         props.i18n ||
         (useI18n({
-          useScope: props.scope as 'global' | 'parent',
-          __useComponent: true
+          useScope: props.scope as 'global' | 'parent'
         }) as unknown as Composer & ComposerInternal)
 
       return renderFormatter<
