--- packages/vue-i18n-core/src/plugin/index.ts
@@ -1,1 +1,0 @@
-export * from './types'
