+++ packages/vue-i18n-core/src/symbols.ts
@@ -6,8 +6,4 @@ export const DatetimePartsSymbol: symbol = /* #__PURE__*/ makeSymbol('__datetime
 export const NumberPartsSymbol: symbol = /* #__PURE__*/ makeSymbol('__numberParts')
 export const EnableEmitter: symbol = /* #__PURE__*/ makeSymbol('__enableEmitter')
 export const DisableEmitter: symbol = /* #__PURE__*/ makeSymbol('__disableEmitter')
-export const SetPluralRulesSymbol: symbol = makeSymbol('__setPluralRules')
-export const DevToolsMetaSymbol: symbol = makeSymbol('__intlifyMeta')
-export const InejctWithOptionSymbol: symbol = /* #__PURE__*/ makeSymbol('__injectWithOption')
-export const ExtendSymbol: symbol = /* #__PURE__*/ makeSymbol('__extend')
 export const DisposeSymbol: symbol = /* #__PURE__*/ makeSymbol('__dispose')
