+++ packages/vue-i18n-core/test/issues.test.ts
@@ -13,9 +13,8 @@ vi.mock('@intlify/shared', async () => {
 })
 
 import { registerMessageCompiler } from '@intlify/core-base'
-import { defineComponent, nextTick, ref } from 'vue'
+import { defineComponent, getCurrentInstance, nextTick, ref } from 'vue'
 import { createI18n, useI18n } from '../src/i18n'
-import { getCurrentInstance } from '../src/utils'
 import { ast } from './fixtures/ast'
 import { mount } from './helper'
